Move TrackingListener to BlockListener
