+++ src/main/java/xyz/holocons/mc/holoitemsrevamp/HoloItemsRevamp.java
@@ -4,8 +4,8 @@ import com.strangeone101.holoitemsapi.Keys;
 import com.strangeone101.holoitemsapi.enchantment.AnvilListener;
 import com.strangeone101.holoitemsapi.enchantment.EnchantManager;
 import com.strangeone101.holoitemsapi.enchantment.EnchantmentListener;
+import com.strangeone101.holoitemsapi.item.BlockListener;
 import com.strangeone101.holoitemsapi.recipe.CraftListener;
-import com.strangeone101.holoitemsapi.tracking.TrackingListener;
 import com.strangeone101.holoitemsapi.tracking.TrackingManager;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -39,7 +39,7 @@ public final class HoloItemsRevamp extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new EnchantmentListener(), this);
         getServer().getPluginManager().registerEvents(new AnvilListener(this), this);
         getServer().getPluginManager().registerEvents(new CraftListener(this), this);
-        getServer().getPluginManager().registerEvents(new TrackingListener(this), this);
+        getServer().getPluginManager().registerEvents(new BlockListener(this), this);
 
         getCommand("holoitems").setExecutor(new MainCommand(this));
         getLogger().info("HoloItems-Revamped [ON]");
